Parallelizable functions need explicit specification of the number of threads.
This involves adding a set_num_threads() method to each of these functions,
which is in turn passed to any OpenMP or SCRAN_CUSTOM_PARALLEL calls. The aim
is to avoid reliance on global variables, which can cause problems with
unintended nested parallelization in complex applications.

Tests have been fixed to handle the new parallelization arguments. In fact, now
that the number of threads is an explicit argument, I have been able to add
some tests to check that the parallelized results are the same as the serial
computation.

Also, I modified summarize_comparisons to avoid mutating the effects array.
This is safer and avoids false sharing that could degrade parallel perf.