Parallelize the PCA with a custom Eigen-like matrix class. (#21)
This introduces a CustomSparseMatrix class that handles the parallelization of
the multiplication, using OpenMP or the SCRAN_CUSTOM_PARALLEL scheme. All
instances of Eigen::SparseMatrix in the various PCA classes are now replaced
with this custom class.

All PCA classes now have a set_num_threads method that is used to set the
number of threads in the CustomSparseMatrix. We also use a Scope object to set
the global number of threads used by OpenMP within the PCA's run() method; 
this is a fallback to ensure that any parallelized Eigen operations
use the specified number of threads.

The internal irlba::Irlba member in the PCA classes has been moved to a
temporary variable within the PCA's run() method, to make life a bit simpler.
All PCA run() methods have also been turned const for consistency.

All accesses into Eigen vectors/matrices are now done with coeff/coeffRef to
avoid any bounds checking, for efficiency. For MultiBatchPCA, we also realize
the coefficient-wise product into an explicit VectorXd, for better cache
locality and to avoid multiple reevaluation of the expression.

We make sure to use a workspace when extracting from the tatami matrix. Somehow
this had been overlooked in the initial implementation; I guess it didn't matter
too much as we were using the matrix's preferred access dimension anyway.

All tests have been added and a timing executable for PCA was added to the
gallery. Some cursory checks indicate that the parallelization does help (4
threads result in a 40% speedup) and the custom matrix class is a little better
than Eigen's SparseMatrix (20% faster when using 4 threads in both
implementations). Though keep in mind the real benefit lies in the fact that we
can insert parallelization schemes other than OpenMP.