Created a separate header to define common macros.
Most of these macros relate to the specification of the custom
parallelization scheme to be used in scran, and thus the propagation of
that scheme to all dependencies. This should make life easier for scran
client applications, who now just need to set SCRAN_CUSTOM_PARALLEL (and
similar parallelization of relevant dependencies is done automatically).

Make sure that this macro header is included in ALL scran files, for
safety's sake. This avoids ODR-related compilation errors when headers
are included directly. We can't defend against ODR errors when
dependencies are included outside of scran, though.