Expose the pairwise effect sizes computed for marker detection. (#23)
This introduces a new PairwiseEffects class that, well, computes the effect
sizes for the pairwise comparisons between groups of cells. This allows users
to just compute the pairwise effects if the summaries are not of interest.

On a related note, we also add a SummarizeEffects class that performs the
summarization for any given effect size. By feeding the output of
PairwiseEffects into SummarizeEffects, users can basically recapitulate the
behavior of ScoreMarkers, but with much more control. 

Much of the code, tests and documentation for ScoreMarkers has been
redistributed to these two new functions. The Factory code has also been
cleaned up to reduce the amount of spaghetti inheritance.