Improve PCA's memory efficiency by using a two-pass strategy. (#24)
The first pass computes the number of non-zero elements in each row; we can
then allocate the exact size of the compressed sparse vectors for filling in
the second pass. This sacrifices a little bit of speed for improved memory
efficiency as we avoid holding one extra copy of the (sub)matrix.

We mitigate some of the extra compute by parallelizing the iterations over the
full matrix to create the compressed sparse vectors. Method-specific compute
in the various PCA classes is also parallelized, closing #22.

The Irlba C++ library version is also bumped to pull in a memory usage fix.