Avoid extraneous copies during SNN clustering. (#26)
This modifies the output of BuildSNNGraph so that we can just create an igraph
vector view on the edges and weights; the views are then used directly to
construct the igraph graph object. This eliminates an unnecessary copy from the
deque of weighted edges into an igraph vector. For further performance
improvements, the SNN edge weight calculations are now parallelized.

We also removed all of the ClusterSNNGraph methods for building the graph
and/or running the clustering from a data matrix. Users should instead use
BuildSNNGraph first to construct the graph, and then pass said graph to the
ClusterSNNGraph classes to generate the desired clustering. (See the updated
README for a demonstration.) This avoids redefinition of several methods.

The igraph wrapper classes have been moved into their own file and namespace.
As the igraph::Vector class is now strictly internal, its copy and move 
constructors have been removed for simplicity.