Switched projection algorithm to avoid creating intermediate artifacts. (#30)
The previous "average of neighbors in PC space" approach would encourage the
formation of intermediate populations or lines in the embedding space, simply
because common embedding algorithms are capable of separating neighbors in the
embedding; the subsequent projection will then take the average of the separate
neighbors, resulting in lines and odd intermediates. 

The new approach "votes" for a single neighbor, based on the distances between
the test cell and the neighbor's neighbors in the embedding, and then projects
the test cell around the elected neighbor. As the projection is localized to a
single neighbor, this ensures that no intermediates can form, at the cost of
some mild formation of tiny clumps.