Store index of each cell's representative after downsampling. (#31)
This makes it easier to reverse the downsampling process for various inferences
like clustering and dimensionality reduction. In particular, we don't need to
perform another nearest neighbor search to identify each cell's representative.

As a result, we can remove the some of the dedicated extraploation functions.
With the known representatives, the propagation of cluster identities becomes
trivial - just use the assigned cluster of each cell's representative, which
is pretty much as good as a majority opinion anyway.

Similarly, each cell can be projected to the coordinates of its representative.
This is no worse than what we're doing now, given that each cell is currently 
stacked on top of a single location with no jitter.