Ignore zeros by default when centering size factors.
This gives a more sane result when dealing with datasets that have loads
of zero size factors from cells that should have been QC'd out but were
not (for whatever reason). Otherwise, the mean size factor is deflated
by all the zeros and this gives inflated normalized expression values
for the few cells that have non-zero counts.