Improve memory efficiency of simple marker detection. (#33)
This eliminates the construction of the pairwise effect size matrix for
Cohen's, LFCs and delta_detected, thus avoiding quadratic memory usage in
ScoreMarkers.  Instead, the underlying per-group statistics are computed from
one pass through the input matrix, and then we loop through the groups to
compute the effect sizes (and thus summaries) for all comparisons involving
each loop. This ensures that we never need to create the entire pairwise effect
matrix in most cases.

Unfortunately, this can't be easily avoided for the AUCs, because (i) there's
no way to get the AUC from per-group statistics, and (ii) we need all AUCs for
each comparison to perform min-rank aggregation. Thus, there is still some
quadratic memory usage, but at least it's not being performed for all the other
effects. 

PairwiseEffects and SummarizeEffects are still valid functions for
back-compatibility. They also serve as useful references to compare against
ScoreMarkers during testing, and I added some more tests to that end.

Hardened tests for the effect size summarization, and also added tests to
compare the different min-rank summarization functions.