Streamlined test suite for differential analyses. (#35)
This removes redundant tests in higher-level functions for scenarios that have
already been checked in lower-level functions. In particular, we added a test
suite for the internal EffectsCalculator, which eliminates the need to check
consistency across matrix representations in ScoreMarkers and PairwiseEffects. 

We also added more tests that compare against a fixed result in simple
scenarios. This provides more protection against refactoring bugs that might
sneak pass the consistency checks (because both the test and actual
implementation are affected). 

Finally, we modified the compare_almost_equal utility so that it is more
relaxed to better handle numerical precision differences in float comparisons.
We also improved the error reporting from this function.