Overhaul the QC filtering classes for easier use. (#36)
In particular, we want to better distinguish between calculation of the
thresholds and application of the filter. This gives users a chance to
intercept and modify the thresholds if so desired. We achieve this by
refactoring the classes so that they return the Thresholds directly; these
Thresholds have their own filtering methods to obtain the discard vector.

In addition:

- Refactored median/MAD calculation into a separate ComputeMedianMad class.
- Renamed IsOutlier to ChooseOutlierFilters, picking up from ComputeMedianMad.
- Renamed PerCell*QcFilters to Suggest*QcFilters.
- Added methods to convert PerCell*QcMetrics::Results to their Buffers.
- Avoid method reordering in doxygen for nicer overload docs.