Updated to the latest version of igraph (0.10.4).
This involves a few internal changes to deal with new signatures. Of particular
interest is the new 'iterations' argument in the Leiden clustering function,
which saves us from having to implement the iterations manually.

Also fixed a bug with the resolution calculation when modularity is used in
Leiden. This brings us into line with igraph::cluster_leiden.