Removed the redundant CustomSparseMatrix class. (#39)
This class is a historical artifact that became obsolete when the parallel
sparse matrix functionality was migrated to irlba's ParallelSparseMatrix. We
just kept it as a wrapper around the PSM to (i) avoid breaking code and (ii)
provide some testing against the Eigen::SparseMatrix case.

Now that we want to add new features, it's a good time to just remove the CSM
in favor of directly using irlba::PSM. This removes a layer of indirection and
allows us to get rid of a stack of test cases (which just repeats the tests in
irlba, and those are more comprehensive anyway). Note that we also lose the
tests for PCA consistency between irlba::PSM and Eigen::SparseMatrix, but those
are covered by stronger checks between irlba::PSM and Eigen::MatrixXd anyway.