Use workspaces for multiplication in irlba wrapper classes. (#40)
This involves updating to a new version of IRLBA that supports workspaces. The
idea is to allocate memory for intermediate data structures once and re-use
that allocated memory across multiply() calls for IRLBA. Similarly, all irlba
wrappers (for blocked and multi-batch PCA) have been updated to support
workspaces. This should avoid excessive allocations and speed things up.

With these workspaces, we need to update all multiply() methods and calls to
include a workspace argument. This is facilitated by the new wrapped_*
functions from the IRLBA library, which also handle the differences between
irlba wrappers and conventional Eigen matrices. We also incorporate the
new wrapped_realize() function for completeness.