Switched back to R's phyper algorithm for HypergeometricTail.
This is way simpler and I suspect it's more robust to double overflows (as the
summation involves terms that generally decrease from 1, due to the fact that
we're computing the lower tail probability "downwards"). We can't cache anymore,
but the function is so fast that the caching doesn't really help anyway.

As we're now doing a lot more invocations of lfactorial, I switched from
Stirling's approximation to Ramanujan's slightly more complicated expression.
This gives a few more significant digits of accuracy that provides a more
accurate p-value when multiple lfactorial calls are performed for choose().

Renamed a couple of the arguments to put it in a gene set enrichment context.