More trend fitting options to handle low density at high abundances. (#49)
This applies a constant-width window for the smoother, which improves the fit
of the trend at high abundances where there are very few points to guide the
smoother. The idea is to avoid stretching the window to include some percentage
of all observations, instead using a fixed width to promote a more local fit.
We do, however, stretch it a little, just to make sure there's enough points.

This new feature takes advantage of the new other options from the latest
version of the CppWeightedLowess library. Relevant options are exposed at both
the FitTrendVar and ModelGeneVar level.