Regression and weighting can now be used at the same time in PCA. (#52)
MultiBatchPCA has been renamed to MultiBatchPca to comply with PascalCasing.

MultiBatchPca can now optionally operate on residuals, whereby the rotation
vectors are computed from the (possibly weighted) residuals, and the input
matrix is projected onto that space to obtain the PCs. The aim here is to get
rotation vectors that focus on the within-batch variation, such that the top
PCs do not just capture the differences between batches (see #46 for details).
At the same time, we don't want to compute PCs from the residuals themselves,
as compositional differences between batches may yield an unsatisfactory result
after regression. So, we project the original matrix onto the rotation vectors,
with the expectation that further batch correction will be done.

In addition, weighting for block size is now optional in MultiBatchPca. This
means that MultiBatchPca has now three modes of operation; weighting only,
residuals only, and weighting + residuals. In all cases, the options only
affect how the blocks contribute to the calculation of the rotation vectors,
and the original matrix is still projected onto the PC space afterwards.

Similarly, we can now equalize the weights of different batches in ResidualPca
(the renamed class of BlockedPCA). This ensures that larger batches do not
dominate the calculation of rotation vectors, just like in MultiBatchPca. The
difference from MultiBatchPca is that we do not do a later projection step,
because we are actually interested in the PCs computed from the residuals.

In short, this commit generalizes the handling of batches so that all combos of
weighting and residuals can be used in MultiBatchPca and ResidualPca.