Refactor ScoreFeatureSet to use the new PCA functions. (#54)
Now we use SimplePca in the unblocked case and ResidualPca in the multi-block
case, which allows us to get rid of a lot of code. It also uses the weighting
of multiple batches from ResidualPca, rather than using its own custom method
for merging rotation vectors across batches. This new approach is a more
natural way of getting a single rotation vector that considers variation
across all batches at once, rather than averaging rotation vectors (which
is not guaranteed to yield a sensible rotation vector if one of the first
rotation vectors is very different.)