Added block size-dependent weights for the PCA functions. (#55)
This introduces a linear weight for each block that is proportional to its
number of cells... up to a certain cap, after which all blocks get the same
weight. The idea is to downweight very small blocks that don't yield reliable
statistics, but to assign equal weight to blocks that are "large enough".