Linking and discovery of igraph is now a developer responsibility.
As it's not a header-only library, it may have already been compiled on
the developer's machine, in which case we might want to use want to use
that rather than re-fetching it ourselves. This is consistent with how
we treat other dependencies (e.g., zlib, HDF5) in other projects.

This means that we have to fetch and link igraph in the gallery and
tests, but this isn't really observable to consumers, so it's fine.