Protect the centering function from non-finite size factors.
Any non-finite size factors are now ignored when computing the mean for
centering. This allows callers to center first and worry about cleaning
up the non-finite size factors afterwards, especially if you want to
ensure that the non-finite factors do not interfere with the centering.

The output of the CenterSizeFactors::run_* methods have been updated to
return a Results struct that specifies whether zero or non-finite values
are detected among the size factors.